requirements/base.txt
flask>=2.3.0
pywebview>=4.4.0
yt-dlp>=2024.1.0
packaging>=21.0
requests>=2.25.0
curl_cffi>=0.5.0
